[arcane,config] Rework handling of default mesh partitioner
Add 'ArcaneGeometricMeshPartitioner' in the list of available default
mesh partitioner and use it if no other is available.
So now we always have an available mesh partitioner.

Move the '#define' for default mesh partitioner in
'arcane_internal_config.h' instead of 'arcane_core_config.h'. This way we
did not have to recompile everything if the default partitioner change.